Solution "Battle_Ground" for theme oop-principles.
