+++ src/main/java/com/rakovets/course/java/core/practice/oop/principles/battle_groundTest/Archer.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop.principles.battleground;
+package com.rakovets.course.java.core.practice.oop.principles.battle_groundTest;
 
 public class Archer extends Hero {
     private final int DAMAGE = 15;
