+++ src/main/java/com/rakovets/course/java/core/practice/oop/principles/battle_groundTest/Mag.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop.principles.battleground;
+package com.rakovets.course.java.core.practice.oop.principles.battle_groundTest;
 
 public class Mag extends Hero {
     private final int DAMAGE = 10;
