+++ src/main/java/com/rakovets/course/java/core/practice/oop/principles/battle_groundTest/Mortal.java
@@ -1,0 +1,5 @@
+package com.rakovets.course.java.core.practice.oop.principles.battle_groundTest;
+
+public interface Mortal {
+    public boolean isAlive();
+}
