+++ src/main/java/com/rakovets/course/java/core/practice/oop/principles/battle_groundTest/Warrior.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop.principles.battleground;
+package com.rakovets.course.java.core.practice.oop.principles.battle_groundTest;
 
 public class Warrior extends Hero {
     private final int DAMAGE = 5;
