--- src/main/java/com/rakovets/course/java/core/practice/oop/principles/battleground/Mortal.java
@@ -1,5 +1,0 @@
-package com.rakovets.course.java.core.practice.oop.principles.battleground;
-
-public interface Mortal {
-    public boolean isAlive();
-}
