--- src/main/java/com/rakovets/course/java/core/practice/oop/principles/battleground/Zombie.java
@@ -1,17 +1,0 @@
-package com.rakovets.course.java.core.practice.oop.principles.battleground;
-
-public class Zombie extends Enemy {
-
-    public Zombie(int health) {
-        super(health);
-    }
-
-    @Override
-    public void attackEnemy(Hero hero) {
-        hero.takeDamage(-10);
-    }
-
-    public void posion(Hero hero) {
-        hero.takeDamage(-20);
-    }
-}
